# Defer schema validation to improve test performance