--- a/.changeset/bright-rules-shout.md
+++ b/.changeset/bright-rules-shout.md
@@ -1,0 +1,5 @@
+---
+'@backstage/catalog-model': patch
+---
+
+Defer `ajv` compilation of schema validators to improve module-import performance