Many modules export an `ajvCompiledJsonSchemaValidator(...)`, which incurs the `ajv` schema compilation at module-import-time. Many tests depend on these modules transitively, but don't exercise the compiled schema - so, this compilation time is wasted.

On my machine, with an example test (airbrake/src/index.test.ts) I'm seeing the following numbers (n=10):

Before: 6005.1ms
After: 5807.8ms
Benefit: 197.3ms (~3.3%)

Weirdly, the NodeJS profiler was saying that schema compilation was taking ~2000ms, but that is likely due to the inaccuracy of the sampling.

#### :heavy_check_mark: Checklist

<!--- Please include the following in your Pull Request when applicable: -->

- [x] A changeset describing the change and affected packages. ([more info](https://github.com/backstage/backstage/blob/master/CONTRIBUTING.md#creating-changesets))
- [x] Added or updated documentation
- [x] Tests for new functionality and regression tests for bug fixes
- [x] Screenshots attached (for UI changes)
- [x] All your commits have a `Signed-off-by` line in the message. ([more info](https://github.com/backstage/backstage/blob/master/CONTRIBUTING.md#developer-certificate-of-origin))